Switch to 24 byte node IDs
This will allow us to accomodate leaf nodes bigger than 4GB, which is
conceivable if a tree is really poorly dimensioned. It will increase the
size of inner nodes by an additional 512 bytes but that will probably be
swamped by the volume of stats.

This change will break any existing data directories.

+++ nodestore/nodestore_test.go
@@ -38,7 +38,7 @@ func TestNodestore(t *testing.T) {
 		bytes := node.ToBytes()
 		_, err = buf.Write(bytes)
 		require.NoError(t, err)
-		addr := nodestore.NewNodeID(1, n, len(bytes))
+		addr := nodestore.NewNodeID(1, uint64(n), uint64(len(bytes)))
 
 		require.NoError(t, ns.Put(ctx, 1, buf.Bytes()))
